Merged PR 12: End to End Containerization
This is a huge PR - sorry. This integrates the API (and DynamoDB) into the full end to end recipe app.
What's working now is: saving a recipe from the UI, retrieving the whole set of saved recipes into the UI.
What's not working: basically everything else.

But you can run the entire stack now with docker-compose, and I've updated the config as well such that running locally won't interfere with how it needs to be deployed.

Related work items: #45